add link subcommand to only handle linking elements
