+++ setup.py
@@ -98,6 +98,7 @@ setup(
             'demo = chanjo.demo:demo',
             'init = chanjo.cli:init_command',
             'load = chanjo.cli:load_command',
+            'link = chanjo.cli:link_command',
             'calculate = chanjo.cli:calculate_command'
         ]
     },
